MainPi/MainStuff/requirements.txt
paho-mqtt==1.6.1
picamera==1.13
pyttsx3==2.90
face-recognition==1.3.0
adafruit-blinka==8.19.0
adafruit-circuitpython-rgb-display==3.11.2
Pillow==9.5.0